[FIRRTL] XMR GCT Interfaces, use correct File/Dir
Generate cross-module references (XMRs) to Grand Central interfaces.
This results in a new "mappings" module that is instantiated in the
companion and contains XMRs to drive the instantiate interface.

Read directory/bind filename info from ExtractGrandCentralAnnotation and
use this to set the correct file and directory names for collateral
generated by Grand Central.

Require that exactly one ExtractGrandCentralAnnotation exists if any
interfaces are encoded on the circuit using AugmentedBundleTypes.  Error
if no annotation is found or if more than one is found.

Add tests of this behavior and erroneous annotation/circuit
combinations.  Fixup all tests to work (since this change places new
requirements on what annotations must exist in order for Grand Central
to not fail).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>