[FIRRTL] Improved Grand Central Anno Scattering
Change Grand Central annotation scattering to represent interfaces with
a recursive structure.  Interfaces which instantiate other interfaces
will show up as BundleTypes in the elements of where they are
instantiated.  This keeps the interface representation intact until the
Grand Central pass can choose to handle this however it wants.  E.g., it
can generate multiple interfaces or it could generate something else.

Change to using numeric IDs instead of relying on string matching to
keep track of interface membership.

This does not update the Grand Central pass to work with this, yet.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>